# Fast skipping on uncompressible data