@stevvooe 2016-01-07
@klauspost You are everywhere!

Any chance of this getting in or should I use the @klauspost fork?