@nigeltao 2016-01-15
The performance deltas are interesting, but I'd actually like the Go snappy implementation to match the C++ snappy implementation in terms of byte-for-byte identical output, when encoding. (Decoding obviously matches byte-for-byte). I don't think it matches now (possibly due to an off by one, maybe a pre-increment vs post-increment somewhere but I can't remember details), but I'd like to do that in preparation for experimenting with asm implementations, and comparing performance to C++'s optimized codegen. (There'll be fewer variable factors if the outputs are identical).

Long story short, if you can get the github.com/google/snappy version (the canonical, C++ version) to accept this encoding algorithm change first, then I'll take it for Go.

Similarly, I think that the Go benchmark files are an identical set to the C++ identical files (or, at least, identical at the time the Go version was written). Further tests over and above what C++ does is fine, but they shouldn't be in the section marked "copied directly from [the C++]". Again, I think that the best outcome would be if you can convince the C++ snappy project to accept that new benchmark first.

Sorry for the late reply. I've been away from work for a few months, for non-work-related reasons.