@klauspost 2016-01-15
Hi Nigel, I hope you had a good leave.

>preparation for experimenting with asm implementations, and comparing performance to C++'s optimized codegen

Sorry, but I don't see why you cannot have your personal research version in a separate branch, and let the users of the package benefit from this clear improvement?