@klauspost 2016-01-15
@nigeltao - Snappy has [similar functionality](https://github.com/google/snappy/blob/master/snappy.cc#L385). They even chose the same skipping sensitivity.

I did find a way to eliminate the extra variable. New benchmark:
```
c:\GoPath\src\github.com\golang\snappy>benchcmp old.txt new.txt
benchmark               old ns/op     new ns/op     delta
Benchmark_ZFlat0-4      386453        395605        +2.37%
Benchmark_ZFlat1-4      5243380       5410402       +3.19%
Benchmark_ZFlat2-4      1219780       18002         -98.52%
Benchmark_ZFlat3-4      1219781       17992         -98.52%
Benchmark_ZFlat4-4      876567        51135         -94.17%
Benchmark_ZFlat5-4      1511665       1562116       +3.34%
Benchmark_ZFlat6-4      1297856       1325112       +2.10%
Benchmark_ZFlat7-4      1134592       1248389       +10.03%
Benchmark_ZFlat8-4      3508325       3586829       +2.24%
Benchmark_ZFlat9-4      4529655       4609800       +1.77%
Benchmark_ZFlat10-4     392659        404117        +2.92%
Benchmark_ZFlat11-4     1060197       1084780       +2.32%
Benchmark_ZFlat12-4     10637050      141784        -98.67%

benchmark               old MB/s     new MB/s     speedup
Benchmark_ZFlat0-4      264.97       258.84       0.98x
Benchmark_ZFlat1-4      133.90       129.77       0.97x
Benchmark_ZFlat2-4      100.91       6837.53      67.76x
Benchmark_ZFlat3-4      100.91       6841.41      67.80x
Benchmark_ZFlat4-4      116.82       2002.54      17.14x
Benchmark_ZFlat5-4      270.96       262.21       0.97x
Benchmark_ZFlat6-4      117.18       114.77       0.98x
Benchmark_ZFlat7-4      110.33       100.27       0.91x
Benchmark_ZFlat8-4      121.64       118.98       0.98x
Benchmark_ZFlat9-4      106.38       104.53       0.98x
Benchmark_ZFlat10-4     302.01       293.45       0.97x
Benchmark_ZFlat11-4     173.85       169.91       0.98x
Benchmark_ZFlat12-4     98.58        7395.55      75.02x
```