@nigeltao 2016-01-16
Ah, I didn't realize that C++ does the same skipping thing. (It's been a while since I read the C++ code). This algorithm change is fine, then.