@klauspost 2016-01-16
I changed the generation, and squashed the commits into one.