@nigeltao 2016-02-07
Ah, sorry, last I looked, I thought that the ball was in your court. I'm not confident that I've been piloting the github UI well, but in one of the diff comments above, I said that:

In any case, as I said, further tests over and above what C++ does is fine, but they shouldn't be in the section marked "copied directly from [the C++]".

The C++ code has had a benchmark named BM_UFlat12 in the past and it might again in the future. I don't want the new Go code to squat on that name. To save more latency on the code review back-and-forth, I just made a git commit to add BenchmarkRandomEncode. The rest of this PR is getting pretty close, bar a few minor nits.