@klauspost 2016-02-07
I implemented the changes and squashed the commits.