@nigeltao 2016-02-08
Sorry, I thought that asking to delete those lines were obvious from what I previously said about "copied directly from [the C++]".