@klauspost 2016-02-09
I renamed the benchmarks to avoid the "confusion".

This is my last change, if you want others, do them yourself, I am not wasting more time on this.