@nigeltao 2016-02-10
I'm sorry that you have found this frustrating. I have also found it frustrating. snappy_test.go contains this declaration:

// testFiles' values are copied directly from
// https://raw.githubusercontent.com/google/snappy/master/snappy_unittest.cc
// etc
var testFiles = etc

I have repeatedly said that I do not want modifications to the testFiles value as it is marked "copied directly from [the C++]". For whatever reason (which may be because I'm not using github correctly; I use it only occasionally compared to gerrit), my request has been ignored.

That's not to say that I'm not interested in benchmarking random input. I did say a few days ago that "I just made a git commit to add BenchmarkRandomEncode", but this also seems to have been ignored, in that your latest version still adds a BenchmarkEncodeRandom function.

In any, case, you have closed this pull request, and said that you are not wasting any more time on this. So be it.

Once again, sorry that you have found this frustrating.